last part is ready

+++ src/main/java/az/edu/itbrains/ecommerce/dtos/basket/BasketAddDto.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BasketAddDto {
-    private Float quantity;
+    private Double quantity;
     private Long productId;
 }
